--- engine/Assets/Scripts/UI/Bars.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: c7ea0048f4d8b4294a7f88cb90a0e23b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
