+++ engine/engine.sln
@@ -1,30 +1,23 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoreEngine", "CoreEngine.csproj", "{660E7428-284B-EED1-7FFD-CA276D12197B}"
+Microsoft Visual Studio Solution File, Format Version 11.00
+# Visual Studio 2010
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoreEngine", "CoreEngine.csproj", "{28740e66-4b28-d1ee-7ffd-ca276d12197b}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlayModeTests", "PlayModeTests.csproj", "{41E8967A-1AD4-6DB7-BA5E-7E9F67B26716}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlayModeTests", "PlayModeTests.csproj", "{7a96e841-d41a-b76d-ba5e-7e9f67b26716}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EditModeTests", "EditModeTests.csproj", "{BB75AA03-C0FB-6074-E634-7AFE38EC939B}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EditModeTests", "EditModeTests.csproj", "{03aa75bb-fbc0-7460-e634-7afe38ec939b}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{660E7428-284B-EED1-7FFD-CA276D12197B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{660E7428-284B-EED1-7FFD-CA276D12197B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{660E7428-284B-EED1-7FFD-CA276D12197B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{660E7428-284B-EED1-7FFD-CA276D12197B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{41E8967A-1AD4-6DB7-BA5E-7E9F67B26716}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{41E8967A-1AD4-6DB7-BA5E-7E9F67B26716}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{41E8967A-1AD4-6DB7-BA5E-7E9F67B26716}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{41E8967A-1AD4-6DB7-BA5E-7E9F67B26716}.Release|Any CPU.Build.0 = Release|Any CPU
-		{BB75AA03-C0FB-6074-E634-7AFE38EC939B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{BB75AA03-C0FB-6074-E634-7AFE38EC939B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{BB75AA03-C0FB-6074-E634-7AFE38EC939B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{BB75AA03-C0FB-6074-E634-7AFE38EC939B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{28740e66-4b28-d1ee-7ffd-ca276d12197b}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{28740e66-4b28-d1ee-7ffd-ca276d12197b}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7a96e841-d41a-b76d-ba5e-7e9f67b26716}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7a96e841-d41a-b76d-ba5e-7e9f67b26716}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{03aa75bb-fbc0-7460-e634-7afe38ec939b}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{03aa75bb-fbc0-7460-e634-7afe38ec939b}.Debug|Any CPU.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
